Refacoring to move stuff stolen from symfony to a separate class (#14)
* Refacoring to move stuff stolen from symfony to a separate class

* Better support for sf2.7

* Applied changes from StyleCI

* Added test for xliff 1.2

* SF2.7 fix

* SF2.7 fix

* Do not skip tests

* Applied changes from StyleCI
