+++ src/Dumper/XliffDumper.php
@@ -11,19 +11,27 @@
 
 namespace Translation\SymfonyStorage\Dumper;
 
-use Symfony\Component\Translation\Dumper\FileDumper;
+use Nyholm\NSA;
+use Symfony\Component\Translation\Dumper\XliffFileDumper;
 use Symfony\Component\Translation\MessageCatalogue;
 use Symfony\Component\Translation\Exception\InvalidArgumentException;
+use Translation\SymfonyStorage\Dumper\Port\SymfonyPort;
 
 /**
  * XliffFileDumper generates xliff files from a message catalogue.
- * Mostly borrowed from Symfony.
+ *
+ * This class provides support for both SF2.7 and SF3.0+
  *
  * @author Tobias Nyholm <tobias.nyholm@gmail.com>
  * @author Michel Salib <michelsalib@hotmail.com>
  */
-final class XliffDumper extends FileDumper
+final class XliffDumper extends XliffFileDumper
 {
+    /**
+     * @var SymfonyPort|null
+     */
+    private $sfPort;
+
     /**
      * {@inheritdoc}
      */
@@ -41,188 +49,24 @@ final class XliffDumper extends FileDumper
         }
 
         if ('1.2' === $xliffVersion) {
-            return $this->dumpXliff1($defaultLocale, $messages, $domain, $options);
+            if (method_exists($this, 'dumpXliff1')) {
+                return NSA::invokeMethod($this, 'dumpXliff1', $defaultLocale, $messages, $domain, $options);
+            } else {
+                // Symfony 2.7
+                return $this->format($messages, $domain);
+            }
         }
+
         if ('2.0' === $xliffVersion) {
-            return $this->dumpXliff2($defaultLocale, $messages, $domain, $options);
+            if (null === $this->sfPort) {
+                $this->sfPort = new SymfonyPort();
+            }
+
+            return $this->sfPort->dumpXliff2($defaultLocale, $messages, $domain, $options);
         }
 
         throw new InvalidArgumentException(
             sprintf('No support implemented for dumping XLIFF version "%s".', $xliffVersion)
         );
-    }
-
-    /**
-     * Symfony 2.7 support.
-     *
-     * @param MessageCatalogue $messages
-     * @param string           $domain
-     *
-     * @return string
-     */
-    protected function format(MessageCatalogue $messages, $domain)
-    {
-        return $this->dumpXliff2(\Locale::getDefault(), $messages, $domain);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    protected function getExtension()
-    {
-        return 'xlf';
-    }
-
-    private function dumpXliff1($defaultLocale, MessageCatalogue $messages, $domain, array $options = [])
-    {
-        $toolInfo = ['tool-id' => 'symfony', 'tool-name' => 'Symfony'];
-        if (array_key_exists('tool_info', $options)) {
-            $toolInfo = array_merge($toolInfo, $options['tool_info']);
-        }
-
-        $dom = new \DOMDocument('1.0', 'utf-8');
-        $dom->formatOutput = true;
-
-        $xliff = $dom->appendChild($dom->createElement('xliff'));
-        $xliff->setAttribute('version', '1.2');
-        $xliff->setAttribute('xmlns', 'urn:oasis:names:tc:xliff:document:1.2');
-
-        $xliffFile = $xliff->appendChild($dom->createElement('file'));
-        $xliffFile->setAttribute('source-language', str_replace('_', '-', $defaultLocale));
-        $xliffFile->setAttribute('target-language', str_replace('_', '-', $messages->getLocale()));
-        $xliffFile->setAttribute('datatype', 'plaintext');
-        $xliffFile->setAttribute('original', 'file.ext');
-
-        $xliffHead = $xliffFile->appendChild($dom->createElement('header'));
-        $xliffTool = $xliffHead->appendChild($dom->createElement('tool'));
-        foreach ($toolInfo as $id => $value) {
-            $xliffTool->setAttribute($id, $value);
-        }
-
-        $xliffBody = $xliffFile->appendChild($dom->createElement('body'));
-        foreach ($messages->all($domain) as $source => $target) {
-            $translation = $dom->createElement('trans-unit');
-
-            $translation->setAttribute(
-                'id',
-                strtr(substr(base64_encode(hash('sha256', $source, true)), 0, 7), '/+', '._')
-            );
-            $translation->setAttribute('resname', $source);
-
-            $s = $translation->appendChild($dom->createElement('source'));
-            $s->appendChild($dom->createTextNode($source));
-
-            // Does the target contain characters requiring a CDATA section?
-            $text = 1 === preg_match('/[&<>]/', $target) ? $dom->createCDATASection($target) : $dom->createTextNode(
-                $target
-            );
-
-            $targetElement = $dom->createElement('target');
-            $metadata = $messages->getMetadata($source, $domain);
-            if ($this->hasMetadataArrayInfo('target-attributes', $metadata)) {
-                foreach ($metadata['target-attributes'] as $name => $value) {
-                    $targetElement->setAttribute($name, $value);
-                }
-            }
-            $t = $translation->appendChild($targetElement);
-            $t->appendChild($text);
-
-            if ($this->hasMetadataArrayInfo('notes', $metadata)) {
-                foreach ($metadata['notes'] as $note) {
-                    if (!isset($note['content'])) {
-                        continue;
-                    }
-
-                    $n = $translation->appendChild($dom->createElement('note'));
-                    $n->appendChild($dom->createTextNode($note['content']));
-
-                    if (isset($note['priority'])) {
-                        $n->setAttribute('priority', $note['priority']);
-                    }
-
-                    if (isset($note['from'])) {
-                        $n->setAttribute('from', $note['from']);
-                    }
-                }
-            }
-
-            $xliffBody->appendChild($translation);
-        }
-
-        return $dom->saveXML();
-    }
-
-    private function dumpXliff2($defaultLocale, MessageCatalogue $messages, $domain, array $options = [])
-    {
-        $dom = new \DOMDocument('1.0', 'utf-8');
-        $dom->formatOutput = true;
-
-        $xliff = $dom->appendChild($dom->createElement('xliff'));
-        $xliff->setAttribute('xmlns', 'urn:oasis:names:tc:xliff:document:2.0');
-        $xliff->setAttribute('version', '2.0');
-        $xliff->setAttribute('srcLang', str_replace('_', '-', $defaultLocale));
-        $xliff->setAttribute('trgLang', str_replace('_', '-', $messages->getLocale()));
-
-        $xliffFile = $xliff->appendChild($dom->createElement('file'));
-        $xliffFile->setAttribute('id', $domain.'.'.$messages->getLocale());
-
-        foreach ($messages->all($domain) as $source => $target) {
-            $translation = $dom->createElement('unit');
-            $translation->setAttribute('id', strtr(substr(base64_encode(hash('sha256', $source, true)), 0, 7), '/+', '._'));
-            $metadata = $messages->getMetadata($source, $domain);
-
-            // Add notes section
-            if ($this->hasMetadataArrayInfo('notes', $metadata)) {
-                $notesElement = $dom->createElement('notes');
-                foreach ($metadata['notes'] as $note) {
-                    $n = $dom->createElement('note');
-                    $n->appendChild($dom->createTextNode(isset($note['content']) ? $note['content'] : ''));
-                    unset($note['content']);
-
-                    foreach ($note as $name => $value) {
-                        $n->setAttribute($name, $value);
-                    }
-                    $notesElement->appendChild($n);
-                }
-                $translation->appendChild($notesElement);
-            }
-
-            $segment = $translation->appendChild($dom->createElement('segment'));
-
-            $s = $segment->appendChild($dom->createElement('source'));
-            $s->appendChild($dom->createTextNode($source));
-
-            // Does the target contain characters requiring a CDATA section?
-            $text = 1 === preg_match('/[&<>]/', $target) ? $dom->createCDATASection($target) : $dom->createTextNode(
-                $target
-            );
-
-            $targetElement = $dom->createElement('target');
-            if ($this->hasMetadataArrayInfo('target-attributes', $metadata)) {
-                foreach ($metadata['target-attributes'] as $name => $value) {
-                    $targetElement->setAttribute($name, $value);
-                }
-            }
-
-            $t = $segment->appendChild($targetElement);
-            $t->appendChild($text);
-
-            $xliffFile->appendChild($translation);
-        }
-
-        return $dom->saveXML();
-    }
-
-    /**
-     * @param string     $key
-     * @param array|null $metadata
-     *
-     * @return bool
-     */
-    private function hasMetadataArrayInfo($key, $metadata = null)
-    {
-        return null !== $metadata &&
-            array_key_exists($key, $metadata) &&
-            ($metadata[$key] instanceof \Traversable || is_array($metadata[$key]));
     }
 }
