+++ tests/Unit/XliffConverterTest.php
@@ -12,7 +12,6 @@
 namespace Translation\SymfonyStorage\Tests\Unit;
 
 use PHPUnit\Framework\TestCase;
-use Symfony\Component\HttpKernel\Kernel;
 use Symfony\Component\Translation\MessageCatalogue;
 use Translation\SymfonyStorage\XliffConverter;
 
@@ -23,10 +22,6 @@ class XliffConverterTest extends TestCase
 {
     public function testContentToCatalogue()
     {
-        if (Kernel::VERSION_ID < 20800) {
-            $this->markTestSkipped('Symfony <2.8 is not supported. ');
-        }
-
         $content = file_get_contents(__DIR__.'/../Fixtures/single-file/messages.en.xlf');
         $catalogue = XliffConverter::contentToCatalogue($content, 'en', 'messages');
